+++ src/components/index.js
@@ -1,7 +1,7 @@
-export { default as ZoaButton } from './inputs/Button.vue';
-export { default as ZoaTextbox } from './inputs/Textbox.vue';
-export { default as ZoaDropdown } from './inputs/Dropdown.vue';
-export { default as ZoaCheckbox } from './inputs/Checkbox.vue';
+export { default as ZoaButton } from './inputs/button/Button.vue';
+export { default as ZoaTextbox } from './inputs/textbox/Textbox.vue';
+export { default as ZoaDropdown } from './inputs/dropdown/Dropdown.vue';
+export { default as ZoaCheckbox } from './inputs/checkbox/Checkbox.vue';
 
-export { default as ZoaFlash } from './dialogs/Flash.vue';
-export { default as ZoaModal } from './dialogs/Modal.vue';
+export { default as ZoaFlash } from './dialogs/flash/Flash.vue';
+export { default as ZoaModal } from './dialogs/modal/Modal.vue';
