+++ src/components/inputs/checkbox/Checkbox.vue
@@ -23,9 +23,9 @@
 </template>
 
 <script setup>
-import { useComponentId } from '../utils/compid.js';
-import FontAwesomeIcon from '../../icons.js';
-import { useChangeEmits } from './common.js';
+import { useComponentId } from '../../utils/compid.js';
+import FontAwesomeIcon from '../../../icons.js';
+import { useChangeEmits } from '../common.js';
 
 const props = defineProps({
   labelPosition: {
@@ -49,7 +49,7 @@ const { valueChanged } = useChangeEmits(emit, props.delay);
 </script>
 
 <style module lang="scss">
-@import 'inputs';
+@import '../inputs';
 
 .grid {
   width: min-content;
