+++ src/components/inputs/dropdown/Dropdown.vue
@@ -26,10 +26,10 @@
 </template>
 
 <script setup>
-import { useComponentId } from '../utils/compid.js';
+import { useComponentId } from '../../utils/compid.js';
 import { computed } from 'vue';
-import FontAwesomeIcon from '../../icons.js';
-import { useChangeEmits } from './common.js';
+import FontAwesomeIcon from '../../../icons.js';
+import { useChangeEmits } from '../common.js';
 
 const props = defineProps({
   labelPosition: {
@@ -63,7 +63,7 @@ const { valueChanged } = useChangeEmits(emit, props.delay);
 </script>
 
 <style module lang="scss">
-@import './inputs';
+@import '../inputs';
 
 .input {
   appearance: none;
